% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bio.dict}
\alias{bio.dict}
\title{Reference table for biomarkers contribution to BHS}
\format{
A reference table to calculate Biological Health Scores (BHS)
\describe{
  \item{Biomarker}{Biomarker names}
  \item{System}{System biomarker belongs to}
  \item{HarmfulInExcess?}{Whether biomarker is considered to be harfmul in excess or not
  Possible values are 1 (Yes), 0 (No), -1 (Not applicable), NA (Not considered)}
}
}
\usage{
bio.dict
}
\description{
A reference table to calculate Biological Health Scores (BHS)
}
\keyword{datasets}
