% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bio.example}
\alias{bio.example}
\title{Example of clean biomarker data frame}
\format{
An object of class \code{data.frame} with 1868 rows and 29 columns.
}
\usage{
bio.example
}
\description{
Toy biomarker dataframe containing measurements for ~2000 individuals (not real data)
}
\keyword{datasets}
