% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bio.original}
\alias{bio.original}
\title{Example of raw biomarker set}
\format{
An object of class \code{data.frame} with 2000 rows and 61 columns.
}
\usage{
bio.original
}
\description{
A dataset containing biomarker measurement with code-like numbers for biomarkers
as well as measurement at two time points
}
\keyword{datasets}
