% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cov}
\alias{cov}
\title{Example of clean covariate data frame}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2000 rows and 23 columns.
}
\usage{
cov
}
\description{
Toy covariate dataframe containing measurements for ~2000 individuals (not real data)
}
\keyword{datasets}
