% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PerformanceSearch.R
\name{factorize}
\alias{factorize}
\title{Function that factorizes character values in a dataframe}
\usage{
factorize(column, df)
}
\description{
It makes dataframes functional for training algorithms
Check if column in integer or string and  turn to the correct type
to avoid other function turning into factor indices
}
