% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BHS.R
\name{formatBio}
\alias{formatBio}
\title{Format and rename colnames of original biomarker dataframe}
\usage{
formatBio(bio.original, measurement = "first")
}
\arguments{
\item{bio.original}{Original biomarker data frame with the first column being ID and the other columns
being biomarker code and some 0 or 1 depending whether the biomarker measurement was the 'at-recruitment' measurement
(0) or the one at follow-up (1)}

\item{measurement}{Measurement time to retrieve. Options: "first"(Default), "second" or "both". Note: if you
choose both the second biomarker colnames will be the biomarker names and ".1" appended to them, this will mess
up with the BHS calculator formatting and is not currently supported. Get in touch if you want to implement that.}
}
\value{
Biomarker data frame with human readable names
}
\description{
The UK Biobank biomarker dataset column names are code names. This function recode those names to match those
of the look up table (bio.dict), so basically readable names.
}
\examples{
library(HDATDS)
data("bio.original_example") # this loads an object called bio.original

bio = formatBio(bio.original)
}
