% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BHS.R
\name{quantile_check}
\alias{quantile_check}
\title{Calculate 1st or 3rd quantile for a given biomarker}
\usage{
quantile_check(column, reference, dataset)
}
\arguments{
\item{column}{biomarker name ["string"] from biomarker data frame, needs to match one name in the
reference look up table}

\item{reference}{Reference "grading" to be taken from the look-up table (possible values: "Mantej",
"Paper", "Barbara")}

\item{dataset}{Biomarker data frame where quantiles will be calculated, this may be the whole data frame
or a subset of it as it is done internally in the BHS calculator when doing stratified BHS calculation}
}
\value{
quartile value for given biomarker, reference and dataset as well as type of quartile nuber (1st or 3rd)
}
\description{
Given a column (biomarker), reference look-up table and dataset (biomarker dataset)
this function calculates the 1st quartile if the given biomarker is considered bad in shortage or
the 3rd quartile if the given biomarker is considered bad in excess
}
